Added Logger placeholder, moved debug code
